+++ angular-frontend-camp/src/app/app-routing.module.ts
@@ -1,7 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { CampListComponent } from './camp-list/tutorials-list.component';
 
-const routes: Routes = [];
+const routes: Routes = [
+{ path: 'camps', component: CampListComponent }
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
