+++ spring-backend/src/main/resources/application.properties
@@ -1,7 +1,20 @@
 # Database connection settings
-spring.datasource.url=jdbc:mysql://localhost:3306/summercamp?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
-spring.datasource.username=summercamp
-spring.datasource.password=summercamp
+#spring.datasource.url=jdbc:mysql://localhost:3306/summercamp?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
+#spring.datasource.url=jdbc:mysql://summercamp-aws.cnpdjzjkgkz1.us-east-2.rds.amazonaws.com
+
+#spring.datasource.username=summercamp
+#spring.datasource.password=Jason-liftoff
+
+#server.port=3306
+
+spring.datasource.url = jdbc:mysql://summercamp-aws.cnpdjzjkgkz1.us-east-2.rds.amazonaws.com:3306/summercamp-aws?useSSL=false&allowPublicKeyRetrieval=true&autoReconnect=true
+spring.datasource.username = summercamp
+spring.datasource.password = Jason-liftoff
+
+
+
+
+
 
 # Specify the DBMS
 spring.jpa.database = MYSQL
